将danmaku_shift字段类型改为double

+++ src/main/java/com/devteam/mikufunbackend/entity/ResourceMatchV0.java
@@ -15,5 +15,5 @@ public class ResourceMatchV0 {
     String episodeTitle;
     String type;
     int episodeId;
-    int shift;
+    double shift;
 }
